GUI/requirements.txt
paho-mqtt==1.6.1
PySide6-Essentials==6.5.0
loguru==0.6.0
pyinstaller==5.6.2
colour==0.1.5
scipy==1.9.0
numpy==1.24.3
bell_avr_libraries-0.1.13-py3-none-any.whl
pyserial